fix(core): conditionally show/label parameters section of execution bar
