+++ app/scripts/modules/core/src/pipeline/executions/execution/execution.less
@@ -127,6 +127,6 @@ execution-details-section-nav {
   }
 }
 
-.execution-details-button {
-  margin-top: 6px;
+.execution-parameters-button {
+  margin-bottom: 6px;
 }
