+++ app/scripts/modules/core/src/pipeline/status/ExecutionStatus.tsx
@@ -19,7 +19,6 @@ import './executionStatus.less';
 export interface IExecutionStatusProps {
   execution: IExecution;
   showingDetails: boolean;
-  showingParams: boolean;
   standalone: boolean;
 }
